Remove unnecessary cast to i32

+++ src/chips/m6502/instructions/load.rs
@@ -7,14 +7,14 @@ impl M6502 {
     }
 
     pub(crate) fn lda(&mut self) {
-        self.a = self.p.set_zero_negative_flags(self.data as i32);
+        self.a = self.p.set_zero_negative_flags(self.data);
     }
 
     pub(crate) fn ldx(&mut self) {
-        self.x = self.p.set_zero_negative_flags(self.data as i32);
+        self.x = self.p.set_zero_negative_flags(self.data);
     }
 
     pub(crate) fn ldy(&mut self) {
-        self.y = self.p.set_zero_negative_flags(self.data as i32);
+        self.y = self.p.set_zero_negative_flags(self.data);
     }
 }
